On suggestion added dto for return types of methods. Added validation for extreme cases. Did Exception handling.

+++ Business/Implementations/DeleteProject.cs
@@ -1,4 +1,5 @@
-﻿using Business.Interface;
+﻿using Business.Execution;
+using Business.Interface;
 using DataAccess;
 using System;
 using System.Collections.Generic;
@@ -10,14 +11,30 @@ namespace Business.Implementations
 {
     public class DeleteProject : IDeleteProject
     {
-        public void Execute(TaskTrackerContext context, int id)
+        public ExecutionResult Execute(TaskTrackerContext context, int id)
         {
+            ExecutionResult exec = new ExecutionResult();
+
             var projectForDeletion = (from projects in context.projects
                                    where projects.Id == id
                                    select projects).FirstOrDefault();
 
-            context.projects.Remove(projectForDeletion);
-            context.SaveChanges();
+            if(projectForDeletion == null)
+            {
+                
+                exec.Error.Add("There is not project with that id");
+                return exec;
+            }
+            else
+            {
+                context.projects.Remove(projectForDeletion);
+                context.SaveChanges();
+                exec.Message.Add("Successfully deleted");
+                return exec;
+            }
+
+
+          
         }
     }
 }
